Calculate pVals with SparkExperimentRunnerComponent

+++ src/main/scala/com/clackjones/connectivitymap/package.scala
@@ -1,7 +1,7 @@
 package com.clackjones
 
 package object connectivitymap {
-  type QuerySignatureMap = Map[String, Int]
+  type QuerySignatureMap = Map[String, Float]
 
   val config = Map(
     "reffileLocation" -> "/home/mike/connectivity_map_resources/reffiles_subset",
